# API_DOCUMENTATION_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Nirium Protocol API
  version: 2.5.0
  description: |
    **Nirium Protocol** - Autonomous AI Agent Matrix on Stellar/Soroban

    This API provides access to:
    - Multi-asset vault management (XLM, USDC, CETES)
    - Autonomous agent orchestration
    - Real-time market data and signals
    - Strategy execution and testing
    - WebSocket subscriptions for live updates

    ## Institutional Security & Compliance (AAA Grade)
    As of Apr 1, 2026, Nirium Protocol adheres to the highest institutional standards:
    - **Data Residency**: Fully compliant with Mexico's LFPDPPP and California's DFAL.
    - **Smart Contract Audits**: Zero critical vulnerabilities. Fuzz-tested with `cargo-fuzz`.
    - **API Security**: 100% compliant with OWASP API Security Top 10. Full-spectrum internal penetration review passed.
    - **JWT**: HS256 algorithm locked — `alg:none` / algorithm-confusion attacks blocked.
    - **Prototype Pollution**: All request bodies stripped of `__proto__`/`constructor`/`prototype` keys before routing.
    - **Host Header Injection**: Incoming `Host` header validated against `nirium.xyz` allowlist on the Edge proxy.
    - **WebSocket**: JWT required via query param (`?token=`). Per-IP connection rate limit: 10 connections/min.
    - **Responsible Disclosure**: `/.well-known/security.txt` (RFC 9116) available at `https://nirium.xyz/.well-known/security.txt`.
...

    ## Authentication

...

...

    2. **API Keys** (long-lived, institutional)
...

...

...
servers:
...
paths:
...
  /api/sandbox/request:
    post:
      tags: [Sandbox]
      summary: Request sandbox account
      description: |
        Request a new sandbox account for institutional testing.
        Returns API key (shown only once) and tier quotas.
      operationId: postSandboxRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SandboxRequest'
      responses:
        '200':
          description: Sandbox account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /api/auth/keys:
    post:
      tags: [Authentication]
      summary: Generate API key
      description: Generate a new API key for the authenticated user
      operationId: postApiKey
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Friendly name for the key
                  example: Production Key
                tier:
                  type: string
                  enum: [free, sandbox, institutional, enterprise]
      responses:
        '200':
          description: API key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  apiKey:
                    type: string
                    description: The API key (shown only once)
                  name:
                    type: string
                  tier:
                    type: string
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [Authentication]
      summary: List API keys
      description: Get all API keys for the authenticated user
      operationId: getApiKeys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/config/llm:
    post:
      tags: [System]
      summary: Configure LLM provider
      description: Update active LLM provider configuration
      operationId: postLlmConfig
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [openai, anthropic, gemini, grok, minimax, bedrock, openrouter, ollama]
                model:
                  type: string
                apiKey:
                  type: string
                ollamaUrl:
                  type: string
      responses:
        '200':
          description: Configuration updated

# ========================================
# COMPONENTS
# ========================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/public/demo-auth or /api/auth/token

    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Long-lived API key obtained from /api/sandbox/request or /api/auth/keys
  schemas:
...
    Quotas:
      type: object
      properties:
        requestsPerMinute:
          type: number
        requestsPerDay:
          type: number
        maxStrategiesPerDay:
          type: number
...
    SandboxRequest:
      type: object
      required: [companyName, contactEmail, walletAddress]
      properties:
        companyName:
          type: string
          example: Institutional Partner LLC
        contactEmail:
          type: string
          format: email
        walletAddress:
          type: string
          pattern: '^G[A-Z0-9]{55}$'
        tier:
          type: string
          enum: [sandbox, institutional]
          default: sandbox
        message:
          type: string
    SandboxAccountResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        account:
          $ref: '#/components/schemas/SandboxAccount'
        warning:
          type: string
    SandboxAccount:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        contactEmail:
          type: string
        walletAddress:
          type: string
        apiKey:
          type: string
          description: Only shown once upon creation
        tier:
          type: string
        quotas:
          $ref: '#/components/schemas/Quotas'
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tier:
          type: string
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Too Many Requests
              retryAfter:
                type: number
                description: Seconds until rate limit resets